requirements.txt
flask==2.3.3
flask-socketio==5.3.6
flask-cors==4.0.0
pymavlink==2.4.37
pyserial==3.5
python-socketio==5.9.0
eventlet==0.33.3
